fix variable size warning

+++ radio/src/gui/colorlcd/model_select.cpp
@@ -741,7 +741,7 @@ class CategoryEditPage : public PageTab
 
         // Details
         char cnt[19];
-        snprintf(cnt, sizeof(cnt), "%u %s", (unsigned)category->size(), STR_MODELS);
+        snprintf(cnt, sizeof(cnt), "%zu %s", category->size(), STR_MODELS);
         new StaticText(window, grid.getFieldSlot(3,1), cnt);
 
         if(category->empty()) {
